+++ Sources/SpotifyWebAPI/API/SpotifyAPI+Personalization.swift
@@ -20,11 +20,13 @@ public extension SpotifyAPI where
      mode. Light or infrequent users of Spotify may not have sufficient
      play history to generate a full affinity data set. As a user’s behavior
      is likely to shift over time, this preference data is available over
-     three time spans. See time_range in the query parameter table for more
+     three time spans. See `timeRange` in the query parameter table for more
      information. For each time range, the top 50 artists are
      available for each user. In the future, it is likely that this
      restriction will be relaxed. This data is typically updated once
      each day for each user.
+     
+     Read more at the [Spotify web API reference][1].
      
      - Parameters:
        - timeRange: *Optional*. Over what time frame the
@@ -33,12 +35,14 @@ public extension SpotifyAPI where
              new data as it becomes available), `mediumTerm` (approximately
              last 6 months), and `shortTerm` (approximately last 4 weeks).
              Default: `mediumTerm`.
-       - offset: Optional. The index of the first artist to return.
+       - offset: *Optional*. The index of the first artist to return.
              Default: 0. Use with limit to get the next set of artists.
        - limit: *Optional*. The number of artists to return.
              Default: 20; Minimum: 1; Maximum: 50.
      - Returns: An array of the full versions of artist objects
            wrapped in a paging object.
+     
+     [1]: https://developer.spotify.com/documentation/web-api/reference/personalization/get-users-top-artists-and-tracks/
      */
     func currentUserTopArtists(
         _ timeRange: TimeRange? = nil,
@@ -78,6 +82,8 @@ public extension SpotifyAPI where
      restriction will be relaxed. This data is typically updated once
      each day for each user.
      
+     Read more at the [Spotify web API reference][1].
+     
      - Parameters:
        - timeRange: *Optional*. Over what time frame the
              affinities are computed. Valid values: `longTerm`
@@ -85,12 +91,14 @@ public extension SpotifyAPI where
              new data as it becomes available), `mediumTerm` (approximately
              last 6 months), and `shortTerm` (approximately last 4 weeks).
              Default: `mediumTerm`.
-       - offset: Optional. The index of the first track to return.
+       - offset: *Optional*. The index of the first track to return.
              Default: 0. Use with limit to get the next set of tracks.
        - limit: *Optional*. The number of tracks to return.
              Default: 20; Minimum: 1; Maximum: 50.
      - Returns: An array of the full versions of track objects
            wrapped in a paging object.
+     
+     [1]: https://developer.spotify.com/documentation/web-api/reference/personalization/get-users-top-artists-and-tracks/
      */
     func currentUserTopTracks(
         _ timeRange: TimeRange? = nil,