+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.1] - 2020-10-24
+
+### Fixed
+
+* Fixed a bug with the JSON data being in the incorrect format for the endpoints for following and unfollowing artists and users.
+* Fixed a typo in the raw value of `TimeRange.shortTerm`, which caused an "invalid request" error from Spotify.
+
 ## [0.8.0] - 2020-10-23
 
 ### Added